chore: refactored save & sign out into its own post controller
