+++ src/main/app/controller/PostController.ts
@@ -2,8 +2,6 @@ import autobind from 'autobind-decorator';
 import { Response } from 'express';
 
 import { getNextStepUrl } from '../../steps';
-import { commonContent } from '../../steps/common/common.content';
-import { saveAndSignOutContent } from '../../steps/save-sign-out/content';
 import { Form } from '../form/Form';
 
 import { AppRequest } from './AppRequest';
@@ -30,12 +28,11 @@ export class PostController<T extends AnyObject> {
       nextUrl = req.url;
     } else {
       await req.locals.api.updateCase(req.session.userCase?.id, formData);
+      if (isSaveAndSignOut) {
+        return;
+      }
       req.session.errors = undefined;
       nextUrl = getNextStepUrl(req, formData);
-    }
-
-    if (isSaveAndSignOut) {
-      return this.saveAndSignOut(req, res);
     }
 
     req.session.save(err => {
@@ -43,23 +40,6 @@ export class PostController<T extends AnyObject> {
         throw err;
       }
       res.redirect(nextUrl);
-    });
-  }
-
-  private saveAndSignOut(req: AppRequest<T>, res: Response) {
-    const email = req.session.user?.email;
-    const language = req.session?.lang || 'en';
-
-    req.session.destroy(err => {
-      if (err) {
-        throw err;
-      }
-
-      res.render(`${__dirname}/../../steps/save-sign-out/template.njk`, {
-        ...commonContent[language],
-        ...saveAndSignOutContent[language],
-        email,
-      });
     });
   }
 }
