+++ src/main/routes.ts
@@ -1,10 +1,12 @@
 import fs from 'fs';
 
-import { Application, RequestHandler } from 'express';
+import { Application, RequestHandler, Response } from 'express';
 
+import { AppRequest } from '../main/app/controller/AppRequest';
 import { GetController } from '../main/app/controller/GetController';
-import { PostController } from '../main/app/controller/PostController';
+import { AnyObject, PostController } from '../main/app/controller/PostController';
 import { Form } from '../main/app/form/Form';
+import { SaveSignOutPostController } from '../main/steps/save-sign-out/post';
 
 import { AccessibilityStatementGetController } from './steps/accessibility-statement/get';
 import { CookiesGetController } from './steps/cookies/get';
@@ -46,7 +48,15 @@ export class Routes {
       app.get(step.url, errorHandler(controller.get));
 
       if (content.form) {
-        app.post(step.url, errorHandler(new PostController(new Form(content.form)).post));
+        const form = new Form(content.form);
+        app.post(
+          step.url,
+          errorHandler((req: AppRequest<AnyObject>, res: Response) =>
+            req.body.saveAndSignOut
+              ? new SaveSignOutPostController(form).post(req, res)
+              : new PostController(form).post(req, res)
+          )
+        );
       }
     }
